Libft_v3

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: mmurakam <mmurakam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/21 19:15:35 by mmurakam          #+#    #+#             */
-/*   Updated: 2020/02/09 03:43:49 by mmurakam         ###   ########.fr       */
+/*   Updated: 2020/05/09 23:18:31 by mmurakam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int count;
